+++ js/components/NotFound/Heading.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export default styled.h2`
+  text-align: center;
+  padding: 5em 0;
+`;
